Moved Testcase in the respective folder
