+++ app/build.gradle
@@ -38,10 +38,10 @@ android {
 }
 
 dependencies {
-
+    testImplementation project(":common:core-testing")
     implementation project(':data')
     implementation project(path: ':domain')
-    testImplementation project(path: ':app')
+
     def worker_hilt="1.0.0"
     def hilt = "2.40.4"
     def hilt_compiler = "2.38.1"
