+++ app/src/test/java/com/example/mvvmcleanarchidemo1/presentation/viewmodel/GameViewModelTest.kt
@@ -2,12 +2,13 @@ package com.example.mvvmcleanarchidemo1.presentation.viewmodel
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.example.mvvmcleanarchidemo1.data.api.FakeNetworkService
+import com.example.core_testing.GameFakeService
+
 import com.example.data.repository.GamesRepoImpl
 import com.example.data.repository.datasourceimpl.RemoteDataSourceImpl
 import com.example.domain.repository.GamesRepo
 import com.example.domain.usecase.GetGamesUseCase
-import com.example.mvvmcleanarchidemo1.data.repository.FakeDataGenerator
+import com.example.core_testing.FakeDataGenerator
 import com.example.mvvmcleanarchidemo1.viewmodels.GameViewModel
 import com.google.common.truth.Truth.assertThat
 import org.junit.Before
@@ -21,13 +22,14 @@ class GameViewModelTest {
     var instantTaskExecutorRule = InstantTaskExecutorRule()
     private lateinit var gamesViewModel: GameViewModel
     private lateinit var fakeUseCase: GetGamesUseCase
-    private lateinit var fakeNetworkService: FakeNetworkService
+    private lateinit var fakeNetworkService: GameFakeService
     private lateinit var accountRepository: GamesRepo
+
 
 
     @Before
     fun setup() {
-        fakeNetworkService = FakeNetworkService(FakeDataGenerator.amiibo)
+        fakeNetworkService = GameFakeService(FakeDataGenerator.amiibo)
         accountRepository = GamesRepoImpl(RemoteDataSourceImpl(fakeNetworkService))
         fakeUseCase = GetGamesUseCase(accountRepository)
         gamesViewModel = GameViewModel(fakeUseCase)
