+++ common/core-testing/src/main/java/com/example/core_testing/FakeDataGenerator.kt
@@ -1,4 +1,4 @@
-package com.example.mvvmcleanarchidemo1.data.repository
+package com.example.core_testing
 
 import com.example.data.models.Amiibo
 import com.example.data.models.GamesAPIResponse
