+++ common/src/main/java/com/example/common/MyClass.kt
@@ -1,0 +1,4 @@
+package com.example.common
+
+class MyClass {
+}
