+++ domain/build.gradle
@@ -33,6 +33,9 @@ android {
 
 dependencies {
 
+
+    testImplementation project(path: ':common:core-testing')
+    testImplementation project(path: ':data')
     def worker_hilt="1.0.0"
     def hilt = "2.40.4"
     def hilt_compiler = "2.38.1"
@@ -49,5 +52,13 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.4'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
+    testImplementation 'org.mockito:mockito-core:3.10.0'
+    testImplementation "org.mockito:mockito-inline:3.10.0"
+    testImplementation 'androidx.arch.core:core-testing:2.1.0'
+    testImplementation"com.squareup.okhttp3:mockwebserver:4.10.0"
+    testImplementation "com.google.truth:truth:1.1"
+    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.1"
+    testImplementation 'org.robolectric:robolectric:4.8'
+    testImplementation 'androidx.test.ext:junit:1.1.4'
 
 }
