+++ settings.gradle
@@ -16,3 +16,5 @@ rootProject.name = "mvvmCleanarchiDemo1"
 include ':app'
 include ':domain'
 include ':data'
+include ':common'
+include ':common:core-testing'
